% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsqlite.R
\name{db.connect}
\alias{db.connect}
\alias{db.open}
\title{Connect to a SQLite database file}
\usage{
db.connect(dbfile, bufsize = NULL)
}
\arguments{
\item{dbfile}{The file path to the database.}

\item{bufsize}{The maximum number of results to read before triggering
a memory allocation. The default is 5000.}
}
\value{
An external pointer that can be used as a database handle
  for the other functions in the package.
}
\description{
Creates a connection with a SQLite database file.
}
