% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsqlite.R
\name{db.disconnect}
\alias{db.disconnect}
\alias{db.close}
\title{Disconnect from a SQLite database file}
\usage{
db.disconnect(db)
}
\arguments{
\item{db}{The database connection handle returned by \code{db.connect}.}
}
\description{
Closes the connection with a SQLite database file.
}
