% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsqlite.R
\name{db.eval}
\alias{db.eval}
\title{Evaluate a SQL statement}
\usage{
db.eval(db, sql)
}
\arguments{
\item{db}{The database connection handle returned by \code{db.connect}.}

\item{sql}{The SQL text to evaluate.}
}
\value{
A matrix representing the result of the evaluation.
}
\description{
Sends a SQL statement to the database and evaluates it.
}
\note{
The matrix is non-standard in the sense that it is a container for
    lists (i.e. subsets of the matrix return lists).
}
