% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsqlite.R
\name{db.info}
\alias{db.info}
\title{Database connection metadata}
\usage{
db.info(db)
}
\arguments{
\item{db}{The database connection handle returned by \code{db.connect}.}
}
\value{
A list. Each item in the list is a token of metadata associated with the
   database connection pointer.
}
\description{
Return the metadata association with a database connection handle.
}
