% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsqlite.R
\name{db.limit}
\alias{db.limit}
\title{Set the result limit}
\usage{
db.limit(db, limit)
}
\arguments{
\item{db}{The database connection handle returned by \code{db.connect}.}

\item{limit}{The new buffer size to use.}
}
\description{
Alters the size on the result buffer.
}
