% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dbBeginTransaction}
\alias{dbBeginTransaction}
\title{Generic for creating a new transaction.}
\usage{
dbBeginTransaction(conn, ...)
}
\arguments{
\item{conn}{An \code{DBIConnection} object.}

\item{...}{Other arguments used by methods}
}
\description{
See method documentation for details.
}

