\docType{methods}
\name{dbClearResult,SQLiteResult-method}
\alias{dbClearResult,SQLiteResult-method}
\alias{sqliteCloseResult}
\title{Clear a result set.}
\usage{
\S4method{dbClearResult}{SQLiteResult}(res, ...)

sqliteCloseResult(res, ...)
}
\arguments{
  \item{res}{an \code{\linkS4class{SQLiteResult}} object.}

  \item{...}{Ignored. Needed for compatibility with
  generic.}
}
\description{
Clear a result set.
}

