% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/table.R
\docType{methods}
\name{dbDataType,SQLiteConnection,ANY-method}
\alias{dbDataType,SQLiteConnection,ANY-method}
\alias{dbDataType,SQLiteDriver,ANY-method}
\title{Determine the SQL Data Type of an R object.}
\usage{
\S4method{dbDataType}{SQLiteConnection,ANY}(dbObj, obj, ...)

\S4method{dbDataType}{SQLiteDriver,ANY}(dbObj, obj, ...)
}
\arguments{
\item{dbObj}{a \code{SQLiteDriver} object,}

\item{obj}{an R object whose SQL type we want to determine.}

\item{...}{Needed for compatibility with generic. Otherwise ignored.}
}
\description{
This method is a straight-forward implementation of the corresponding
generic function.
}
\examples{
dbDataType(SQLite(), 1)
dbDataType(SQLite(), 1L)
dbDataType(SQLite(), "1")
dbDataType(SQLite(), TRUE)
dbDataType(SQLite(), list())

sapply(datasets::quakes, dbDataType, dbObj = SQLite())
}

