% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{dbDisconnect,SQLiteConnection-method}
\alias{dbDisconnect,SQLiteConnection-method}
\title{Disconnect an SQLite connection.}
\usage{
\S4method{dbDisconnect}{SQLiteConnection}(conn)
}
\arguments{
\item{conn}{An existing \code{\linkS4class{SQLiteConnection}}}
}
\description{
Disconnect an SQLite connection.
}

