\name{dbExistsTable,SQLiteConnection,character-method}
\alias{dbExistsTable,SQLiteConnection,character-method}
\title{Does the table exist?}
\usage{
  \S4method{dbExistsTable}{SQLiteConnection,character}(conn,
    name, ...)
}
\arguments{
  \item{conn}{An existing
  \code{\linkS4class{SQLiteConnection}}}

  \item{...}{Ignored. Included for compatibility with
  generic.}
}
\description{
  Does the table exist?
}

