% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Connection.R
\docType{methods}
\name{dbGetException,SQLiteConnection-method}
\alias{dbGetException,SQLiteConnection-method}
\title{Get the last exception from the connection.}
\usage{
\S4method{dbGetException}{SQLiteConnection}(conn)
}
\arguments{
\item{conn}{an object of class \code{\linkS4class{SQLiteConnection}}}
}
\description{
Get the last exception from the connection.
}
\keyword{internal}

