% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{dbListConnections,SQLiteDriver-method}
\alias{dbListConnections,SQLiteDriver-method}
\title{List active connections}
\usage{
\S4method{dbListConnections}{SQLiteDriver}(drv, ...)
}
\arguments{
\item{drv}{An object of class \code{\linkS4class{SQLiteDriver}}}

\item{...}{Ignored. Needed for compatibility with generic.}
}
\description{
List active connections
}

