% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{dbListResults,SQLiteConnection-method}
\alias{dbListResults,SQLiteConnection-method}
\title{List available SQLite result sets.}
\usage{
\S4method{dbListResults}{SQLiteConnection}(conn)
}
\arguments{
\item{conn}{An existing \code{\linkS4class{SQLiteConnection}}}
}
\description{
List available SQLite result sets.
}

