% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Driver.R
\docType{methods}
\name{dbUnloadDriver,SQLiteDriver-method}
\alias{dbUnloadDriver,SQLiteDriver-method}
\title{Unload SQLite driver.}
\usage{
\S4method{dbUnloadDriver}{SQLiteDriver}(drv, ...)
}
\arguments{
\item{drv}{Object created by \code{\link{SQLite}}}

\item{...}{Ignored. Needed for compatibility with generic.}
}
\description{
This function is no longer needed.
}
\keyword{internal}

