% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\docType{methods}
\name{dummy-methods}
\alias{dummy-methods}
\alias{dbGetQuery,NULL,ANY-method}
\title{Dummy methods}
\usage{
\S4method{dbGetQuery}{`NULL`,ANY}(conn, statement, ...)
}
\arguments{
\item{conn}{A \code{\linkS4class{DBIConnection}} object, as produced by
\code{\link[=dbConnect]{dbConnect()}}.}

\item{statement}{a character vector of length 1 containing SQL.}

\item{...}{Other parameters passed on to methods.}
}
\description{
Define here so that these methods can also be imported from this package.
Recommended practice is to import all methods from \pkg{DBI}.
}

