% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{methods}
\name{fetch,SQLiteResult-method}
\alias{fetch,SQLiteResult-method}
\title{Fetch.}
\usage{
\S4method{fetch}{SQLiteResult}(res, n = -1, ..., row.names = NA)
}
\description{
Deprecated. Please use \code{dbFetch} instead.
}
\keyword{internal}

