% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_SQLiteResult.R
\name{fetch_SQLiteResult}
\alias{fetch_SQLiteResult}
\alias{fetch,SQLiteResult-method}
\title{Fetch}
\usage{
\S4method{fetch}{SQLiteResult}(res, n = -1, ...)
}
\description{
A shortcut for \code{\link[DBI]{dbFetch}(res, n = n, row.names = FALSE)},
kept for compatibility reasons.
}
\keyword{internal}
