% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initRegExp.R
\name{initRegExp}
\alias{initRegExp}
\title{Add regular expression operator}
\usage{
initRegExp(db)
}
\arguments{
\item{db}{A \code{\linkS4class{SQLiteConnection}} object to add the regular expression operator into the connection.}
}
\value{
Always \code{TRUE}, invisibly.
}
\description{
Forwarded to \code{initExtension(db, "regexp")} .
}
\keyword{internal}
