% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{initSeries}
\alias{initSeries}
\title{Add \code{generate_series()} function}
\usage{
initSeries(db)
}
\arguments{
\item{db}{A \code{\linkS4class{SQLiteConnection}} object to load these extensions into.}
}
\value{
Always \code{TRUE}, invisibly.
}
\description{
This loads the table-valued function \code{generate_series()},
as available through the SQLite source code repository
(\url{https://sqlite.org/src/raw?filename=ext/misc/series.c}).
}
\examples{
library(DBI)
db <- RSQLite::datasetsDb()
RSQLite::initSeries(db)

dbGetQuery(db, "SELECT value FROM generate_series(0, 20, 5);")
dbDisconnect(db)
}
