% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsqlite.R
\name{list.fields}
\alias{list.fields}
\title{Table fields}
\usage{
list.fields(db, name)
}
\arguments{
\item{db}{The database connection handle returned by \code{db.connect}.}

\item{name}{The name of a database table.}
}
\value{
A vector of field names.
}
\description{
Return the fields associated with a table.
}
