% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsqlite.R
\name{list.tables}
\alias{list.tables}
\title{Database tables}
\usage{
list.tables(db)
}
\arguments{
\item{db}{The database connection handle returned by \code{db.connect}.}
}
\value{
A vector of table names.
}
\description{
Return the names of tables in the database.
}
