% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\docType{import}
\name{reexports}
\alias{dbDriver}
\alias{reexports}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{DBI}{\code{\link[DBI]{dbDriver}}}
}}

