\name{S4R}
\alias{ErrorClass}
\alias{usingR}

\title{R compatibility with S version 4/S-Plus 5+ support functions}
\description{
  These objects ease the task of porting functions into R
  from S Version 4 and S-Plus 5.0 and later.  See the documentation
  of the lower-case functions there. May be obsolete in the future.
}
\usage{
usingR(major, minor)
}
\keyword{internal}
% vim:syntax=tex
