% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLiteConnection.R,
%   R/dbAppendTableArrow_SQLiteConnection.R,
%   R/dbAppendTable_SQLiteConnection.R,
%   R/dbCreateTableArrow_SQLiteConnection.R, R/dbDataType_SQLiteConnection.R,
%   R/dbExistsTable_SQLiteConnection_Id.R,
%   R/dbExistsTable_SQLiteConnection_character.R,
%   R/dbGetException_SQLiteConnection.R, R/dbGetInfo_SQLiteConnection.R,
%   R/dbIsValid_SQLiteConnection.R, R/dbListObjects_SQLiteConnection.R,
%   R/dbListTables_SQLiteConnection.R,
%   R/dbQuoteIdentifier_SQLiteConnection_SQL.R,
%   R/dbQuoteIdentifier_SQLiteConnection_character.R,
%   R/dbRemoveTable_SQLiteConnection_character.R,
%   R/dbSendQueryArrow_SQLiteConnection.R,
%   R/dbSendQuery_SQLiteConnection_character.R,
%   R/dbUnquoteIdentifier_SQLiteConnection_SQL.R,
%   R/dbWriteTableArrow_SQLiteConnection.R, R/show_SQLiteConnection.R,
%   R/sqlData_SQLiteConnection.R
\docType{class}
\name{SQLiteConnection-class}
\alias{SQLiteConnection-class}
\alias{format.SQLiteConnection}
\alias{dbAppendTableArrow_SQLiteConnection}
\alias{dbAppendTableArrow,SQLiteConnection-method}
\alias{dbAppendTable_SQLiteConnection}
\alias{dbAppendTable,SQLiteConnection-method}
\alias{dbCreateTableArrow_SQLiteConnection}
\alias{dbCreateTableArrow,SQLiteConnection-method}
\alias{dbDataType_SQLiteConnection}
\alias{dbDataType,SQLiteConnection-method}
\alias{dbExistsTable_SQLiteConnection_Id}
\alias{dbExistsTable,SQLiteConnection,Id-method}
\alias{dbExistsTable_SQLiteConnection_character}
\alias{dbExistsTable,SQLiteConnection,character-method}
\alias{dbGetException_SQLiteConnection}
\alias{dbGetException,SQLiteConnection-method}
\alias{dbGetInfo_SQLiteConnection}
\alias{dbGetInfo,SQLiteConnection-method}
\alias{dbIsValid_SQLiteConnection}
\alias{dbIsValid,SQLiteConnection-method}
\alias{dbListObjects_SQLiteConnection}
\alias{dbListObjects,SQLiteConnection-method}
\alias{dbListTables_SQLiteConnection}
\alias{dbListTables,SQLiteConnection-method}
\alias{dbQuoteIdentifier_SQLiteConnection_SQL}
\alias{dbQuoteIdentifier,SQLiteConnection,SQL-method}
\alias{dbQuoteIdentifier_SQLiteConnection_character}
\alias{dbQuoteIdentifier,SQLiteConnection,character-method}
\alias{dbRemoveTable_SQLiteConnection_character}
\alias{dbRemoveTable,SQLiteConnection,character-method}
\alias{dbSendQueryArrow_SQLiteConnection}
\alias{dbSendQueryArrow,SQLiteConnection-method}
\alias{dbSendQuery_SQLiteConnection_character}
\alias{dbSendQuery,SQLiteConnection,character-method}
\alias{dbUnquoteIdentifier_SQLiteConnection_SQL}
\alias{dbUnquoteIdentifier,SQLiteConnection,SQL-method}
\alias{dbWriteTableArrow_SQLiteConnection}
\alias{dbWriteTableArrow,SQLiteConnection-method}
\alias{show_SQLiteConnection}
\alias{show,SQLiteConnection-method}
\alias{sqlData_SQLiteConnection}
\alias{sqlData,SQLiteConnection-method}
\title{Class SQLiteConnection (and methods)}
\usage{
\method{format}{SQLiteConnection}(x, ...)

\S4method{dbAppendTableArrow}{SQLiteConnection}(conn, name, value, ...)

\S4method{dbAppendTable}{SQLiteConnection}(conn, name, value, ..., row.names = NULL)

\S4method{dbCreateTableArrow}{SQLiteConnection}(
  conn,
  name,
  value,
  ...,
  field.types = NULL,
  temporary = FALSE
)

\S4method{dbDataType}{SQLiteConnection}(dbObj, obj, ...)

\S4method{dbExistsTable}{SQLiteConnection,Id}(conn, name, ...)

\S4method{dbExistsTable}{SQLiteConnection,character}(conn, name, ...)

\S4method{dbGetException}{SQLiteConnection}(conn, ...)

\S4method{dbGetInfo}{SQLiteConnection}(dbObj, ...)

\S4method{dbIsValid}{SQLiteConnection}(dbObj, ...)

\S4method{dbListObjects}{SQLiteConnection}(conn, prefix = NULL, ...)

\S4method{dbListTables}{SQLiteConnection}(conn, ...)

\S4method{dbQuoteIdentifier}{SQLiteConnection,SQL}(conn, x, ...)

\S4method{dbQuoteIdentifier}{SQLiteConnection,character}(conn, x, ...)

\S4method{dbRemoveTable}{SQLiteConnection,character}(conn, name, ..., temporary = FALSE, fail_if_missing = TRUE)

\S4method{dbSendQueryArrow}{SQLiteConnection}(conn, statement, params = NULL, ...)

\S4method{dbSendQuery}{SQLiteConnection,character}(conn, statement, params = NULL, ...)

\S4method{dbUnquoteIdentifier}{SQLiteConnection,SQL}(conn, x, ...)

\S4method{dbWriteTableArrow}{SQLiteConnection}(
  conn,
  name,
  value,
  append = FALSE,
  overwrite = FALSE,
  ...,
  temporary = FALSE,
  field.types = NULL
)

\S4method{show}{SQLiteConnection}(object)

\S4method{sqlData}{SQLiteConnection}(
  con,
  value,
  row.names = pkgconfig::get_config("RSQLite::row.names.query", FALSE),
  ...
)
}
\arguments{
\item{...}{Other parameters passed on to methods.}

\item{conn}{A \link[DBI:DBIConnection-class]{DBI::DBIConnection} object,
as returned by \code{\link[DBI:dbConnect]{dbConnect()}}.}

\item{name}{The table name, passed on to \code{\link[DBI:dbQuoteIdentifier]{dbQuoteIdentifier()}}. Options are:
\itemize{
\item a character string with the unquoted DBMS table name,
e.g. \code{"table_name"},
\item a call to \code{\link[DBI:Id]{Id()}} with components to the fully qualified table name,
e.g. \code{Id(schema = "my_schema", table = "table_name")}
\item a call to \code{\link[DBI:SQL]{SQL()}} with the quoted and fully qualified table name
given verbatim, e.g. \code{SQL('"my_schema"."table_name"')}
}}

\item{value}{An object coercible with \code{\link[nanoarrow:as_nanoarrow_array_stream]{nanoarrow::as_nanoarrow_array_stream()}}.}

\item{field.types}{character vector of named SQL field types where the names are the names of new table's columns.
If missing, types inferred with \code{\link[DBI:dbDataType]{DBI::dbDataType()}}).}

\item{temporary}{If \code{TRUE}, only temporary tables are considered.}

\item{fail_if_missing}{If \code{FALSE}, \code{dbRemoveTable()} succeeds if the table doesn't exist.}
}
\description{
SQLiteConnection objects are created by passing \code{\link[=SQLite]{SQLite()}} as first argument to \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.
They are a superclass of the \link[DBI:DBIConnection-class]{DBIConnection} class.
The "Usage" section lists the class methods overridden by \pkg{RSQLite}.
}
\seealso{
The corresponding generic functions
\code{\link[DBI:dbSendQuery]{DBI::dbSendQuery()}}, \code{\link[DBI:dbGetQuery]{DBI::dbGetQuery()}},
\code{\link[DBI:dbSendStatement]{DBI::dbSendStatement()}}, \code{\link[DBI:dbExecute]{DBI::dbExecute()}},
\code{\link[DBI:dbExistsTable]{DBI::dbExistsTable()}}, \code{\link[DBI:dbListTables]{DBI::dbListTables()}}, \code{\link[DBI:dbListFields]{DBI::dbListFields()}},
\code{\link[DBI:dbRemoveTable]{DBI::dbRemoveTable()}}, and \code{\link[DBI:sqlData]{DBI::sqlData()}}.
}
\keyword{internal}
