% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLiteResult.R, R/dbBindArrow_SQLiteResult.R,
%   R/dbBind_SQLiteResult.R, R/dbClearResult_SQLiteResult.R,
%   R/dbColumnInfo_SQLiteResult.R, R/dbFetchArrowChunk_SQLiteResult.R,
%   R/dbFetchArrow_SQLiteResult.R, R/dbFetch_SQLiteResult.R,
%   R/dbGetRowCount_SQLiteResult.R, R/dbGetRowsAffected_SQLiteResult.R,
%   R/dbGetStatement_SQLiteResult.R, R/dbHasCompleted_SQLiteResult.R,
%   R/dbIsValid_SQLiteResult.R
\docType{class}
\name{SQLiteResult-class}
\alias{SQLiteResult-class}
\alias{dbBindArrow_SQLiteResult}
\alias{dbBindArrow,SQLiteResult-method}
\alias{dbBind_SQLiteResult}
\alias{dbBind,SQLiteResult-method}
\alias{dbClearResult_SQLiteResult}
\alias{dbClearResult,SQLiteResult-method}
\alias{dbColumnInfo_SQLiteResult}
\alias{dbColumnInfo,SQLiteResult-method}
\alias{dbFetchArrowChunk_SQLiteResult}
\alias{dbFetchArrowChunk,SQLiteResult-method}
\alias{dbFetchArrow_SQLiteResult}
\alias{dbFetchArrow,SQLiteResult-method}
\alias{dbFetch_SQLiteResult}
\alias{dbFetch,SQLiteResult-method}
\alias{dbGetRowCount_SQLiteResult}
\alias{dbGetRowCount,SQLiteResult-method}
\alias{dbGetRowsAffected_SQLiteResult}
\alias{dbGetRowsAffected,SQLiteResult-method}
\alias{dbGetStatement_SQLiteResult}
\alias{dbGetStatement,SQLiteResult-method}
\alias{dbHasCompleted_SQLiteResult}
\alias{dbHasCompleted,SQLiteResult-method}
\alias{dbIsValid_SQLiteResult}
\alias{dbIsValid,SQLiteResult-method}
\title{Class SQLiteResult (and methods)}
\usage{
\S4method{dbBindArrow}{SQLiteResult}(res, params, ...)

\S4method{dbBind}{SQLiteResult}(res, params, ...)

\S4method{dbClearResult}{SQLiteResult}(res, ...)

\S4method{dbColumnInfo}{SQLiteResult}(res, ...)

\S4method{dbFetchArrowChunk}{SQLiteResult}(res, ..., chunk_size = 65536L)

\S4method{dbFetchArrow}{SQLiteResult}(res, ..., chunk_size = 65536L)

\S4method{dbFetch}{SQLiteResult}(
  res,
  n = -1,
  ...,
  row.names = pkgconfig::get_config("RSQLite::row.names.query", FALSE)
)

\S4method{dbGetRowCount}{SQLiteResult}(res, ...)

\S4method{dbGetRowsAffected}{SQLiteResult}(res, ...)

\S4method{dbGetStatement}{SQLiteResult}(res, ...)

\S4method{dbHasCompleted}{SQLiteResult}(res, ...)

\S4method{dbIsValid}{SQLiteResult}(dbObj, ...)
}
\arguments{
\item{res}{An object inheriting from \link[DBI:DBIResult-class]{DBI::DBIResult}.}

\item{params}{For \code{dbBind()}, a list of values, named or unnamed,
or a data frame, with one element/column per query parameter.
For \code{dbBindArrow()}, values as a nanoarrow stream,
with one column per query parameter.}

\item{...}{Other arguments passed on to methods.}

\item{chunk_size}{The number of rows in each Arrow array of the stream returned by \code{dbFetchArrow()},
or in the array returned by \code{dbFetchArrowChunk()}, see \link{sqlite-arrow}.}
}
\description{
SQLiteDriver objects are created by \code{\link[DBI:dbSendQuery]{dbSendQuery()}} or \code{\link[DBI:dbSendStatement]{dbSendStatement()}},
and encapsulate the result of an SQL statement (either \code{SELECT} or not).
They are a superclass of the \link[DBI:DBIResult-class]{DBIResult} class.
The "Usage" section lists the class methods overridden by \pkg{RSQLite}.
}
\seealso{
The corresponding generic functions
\code{\link[DBI:dbFetch]{DBI::dbFetch()}}, \code{\link[DBI:dbClearResult]{DBI::dbClearResult()}}, and \code{\link[DBI:dbBind]{DBI::dbBind()}},
\code{\link[DBI:dbColumnInfo]{DBI::dbColumnInfo()}}, \code{\link[DBI:dbGetRowsAffected]{DBI::dbGetRowsAffected()}}, \code{\link[DBI:dbGetRowCount]{DBI::dbGetRowCount()}},
\code{\link[DBI:dbHasCompleted]{DBI::dbHasCompleted()}}, and \code{\link[DBI:dbGetStatement]{DBI::dbGetStatement()}}.
}
\keyword{internal}
