% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLiteResultArrow.R,
%   R/dbBindArrow_SQLiteResultArrow.R, R/dbBind_SQLiteResultArrow.R,
%   R/dbClearResult_SQLiteResultArrow.R, R/dbColumnInfo_SQLiteResultArrow.R,
%   R/dbFetchArrowChunk_SQLiteResultArrow.R,
%   R/dbFetchArrow_SQLiteResultArrow.R, R/dbFetch_SQLiteResultArrow.R,
%   R/dbGetRowCount_SQLiteResultArrow.R,
%   R/dbGetRowsAffected_SQLiteResultArrow.R,
%   R/dbGetStatement_SQLiteResultArrow.R, R/dbHasCompleted_SQLiteResultArrow.R,
%   R/dbIsValid_SQLiteResultArrow.R, R/show_SQLiteResultArrow.R
\docType{class}
\name{SQLiteResultArrow-class}
\alias{SQLiteResultArrow-class}
\alias{dbBindArrow_SQLiteResultArrow}
\alias{dbBindArrow,SQLiteResultArrow-method}
\alias{dbBind_SQLiteResultArrow}
\alias{dbBind,SQLiteResultArrow-method}
\alias{dbClearResult_SQLiteResultArrow}
\alias{dbClearResult,SQLiteResultArrow-method}
\alias{dbColumnInfo_SQLiteResultArrow}
\alias{dbColumnInfo,SQLiteResultArrow-method}
\alias{dbFetchArrowChunk_SQLiteResultArrow}
\alias{dbFetchArrowChunk,SQLiteResultArrow-method}
\alias{dbFetchArrow_SQLiteResultArrow}
\alias{dbFetchArrow,SQLiteResultArrow-method}
\alias{dbFetch_SQLiteResultArrow}
\alias{dbFetch,SQLiteResultArrow-method}
\alias{dbGetRowCount_SQLiteResultArrow}
\alias{dbGetRowCount,SQLiteResultArrow-method}
\alias{dbGetRowsAffected_SQLiteResultArrow}
\alias{dbGetRowsAffected,SQLiteResultArrow-method}
\alias{dbGetStatement_SQLiteResultArrow}
\alias{dbGetStatement,SQLiteResultArrow-method}
\alias{dbHasCompleted_SQLiteResultArrow}
\alias{dbHasCompleted,SQLiteResultArrow-method}
\alias{dbIsValid_SQLiteResultArrow}
\alias{dbIsValid,SQLiteResultArrow-method}
\alias{show_SQLiteResultArrow}
\alias{show,SQLiteResultArrow-method}
\title{Class SQLiteResultArrow (and methods)}
\usage{
\S4method{dbBindArrow}{SQLiteResultArrow}(res, params, ...)

\S4method{dbBind}{SQLiteResultArrow}(res, params, ...)

\S4method{dbClearResult}{SQLiteResultArrow}(res, ...)

\S4method{dbColumnInfo}{SQLiteResultArrow}(res, ...)

\S4method{dbFetchArrowChunk}{SQLiteResultArrow}(res, ..., chunk_size = 65536L)

\S4method{dbFetchArrow}{SQLiteResultArrow}(res, ..., chunk_size = 65536L)

\S4method{dbFetch}{SQLiteResultArrow}(res, n = -1, ...)

\S4method{dbGetRowCount}{SQLiteResultArrow}(res, ...)

\S4method{dbGetRowsAffected}{SQLiteResultArrow}(res, ...)

\S4method{dbGetStatement}{SQLiteResultArrow}(res, ...)

\S4method{dbHasCompleted}{SQLiteResultArrow}(res, ...)

\S4method{dbIsValid}{SQLiteResultArrow}(dbObj, ...)

\S4method{show}{SQLiteResultArrow}(object)
}
\arguments{
\item{res}{An object inheriting from \link[DBI:DBIResult-class]{DBI::DBIResult}.}

\item{params}{For \code{dbBind()}, a list of values, named or unnamed,
or a data frame, with one element/column per query parameter.
For \code{dbBindArrow()}, values as a nanoarrow stream,
with one column per query parameter.}

\item{...}{Other arguments passed on to methods.}

\item{chunk_size}{The number of rows in each Arrow array of the stream returned by \code{dbFetchArrow()},
or in the array returned by \code{dbFetchArrowChunk()}, see \link{sqlite-arrow}.}

\item{n}{maximum number of records to retrieve per fetch. Use \code{n = -1}
or \code{n = Inf}
to retrieve all pending records.  Some implementations may recognize other
special values.}

\item{dbObj}{An object inheriting from \link[DBI:DBIObject-class]{DBIObject},
i.e. \link[DBI:DBIDriver-class]{DBIDriver}, \link[DBI:DBIConnection-class]{DBIConnection},
or a \link[DBI:DBIResult-class]{DBIResult}}
}
\description{
SQLiteResultArrow objects are created by \code{\link[DBI:dbSendQueryArrow]{DBI::dbSendQueryArrow()}},
and encapsulate the result of an SQL query whose rows are fetched as Arrow arrays.
They are a superclass of the \link[DBI:DBIResultArrow-class]{DBIResultArrow-class} class.
The "Usage" section lists the class methods overridden by \pkg{RSQLite}.
Every other method delegates to the \link[=SQLiteResult-class]{SQLiteResult} object in the \code{result} slot.
}
\seealso{
The corresponding generic functions
\code{\link[DBI:dbFetchArrow]{DBI::dbFetchArrow()}}, \code{\link[DBI:dbFetchArrowChunk]{DBI::dbFetchArrowChunk()}}, \code{\link[DBI:dbBindArrow]{DBI::dbBindArrow()}},
\code{\link[DBI:dbClearResult]{DBI::dbClearResult()}}, and \code{\link[DBI:dbHasCompleted]{DBI::dbHasCompleted()}},
and the type mappings in \link{sqlite-arrow}.
}
\keyword{internal}
