% Generated by roxygen2 (4.0.1): do not edit by hand
\name{usingR}
\alias{ErrorClass}
\alias{usingR}
\title{R compatibility with S version 4/S-Plus 5+ support functions}
\usage{
usingR(major = 0, minor = 0)
}
\description{
These objects ease the task of porting functions into R from S Version 4 and
S-Plus 5.0 and later.  See the documentation of the lower-case functions
there. May be obsolete in the future.
}
\keyword{internal}

