% Generated by roxygen2 (4.0.2): do not edit by hand
\name{datasetsDb}
\alias{datasetsDb}
\title{A sample sqlite database.}
\usage{
datasetsDb()
}
\description{
This database is bundled with the package, and contains all data frames
in the datasets package.
}
\examples{
db <- datasetsDb()
dbListTables(db)

dbReadTable(db, "CO2")
dbGetQuery(db, "SELECT * FROM CO2 WHERE conc < 100")

dbDisconnect(db)
}

