\name{dbBeginTransaction}
\alias{dbBeginTransaction}
\title{Generic for creating a new transaction.}
\usage{
dbBeginTransaction(conn, ...)
}
\arguments{
  \item{conn}{An \code{DBIConnection} object.}

  \item{...}{Other arguments used by methods}
}
\description{
See method documentation for details.
}

