% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{dbBeginTransaction}
\alias{dbBeginTransaction}
\title{Generic for creating a new transaction}
\usage{
dbBeginTransaction(conn, ...)
}
\arguments{
\item{conn}{A \code{DBIConnection} object.}

\item{...}{Other arguments used by methods}
}
\description{
See method documentation for details.
}
\keyword{internal}
