% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{dbCallProc,SQLiteConnection-method}
\alias{dbCallProc,SQLiteConnection-method}
\title{Call an SQL stored procedure}
\usage{
\S4method{dbCallProc}{SQLiteConnection}(conn, ...)
}
\arguments{
\item{conn}{
  a \code{DBIConnection} object.
}

\item{...}{
  additional arguments are passed to the implementing method.
}
}
\description{
Not yet implemented.
}
\keyword{internal}

