\docType{methods}
\name{dbCallProc,SQLiteConnection-method}
\alias{dbCallProc,SQLiteConnection-method}
\title{Call an SQL stored procedure}
\usage{
\S4method{dbCallProc}{SQLiteConnection}(conn, ...)
}
\arguments{
  \item{conn}{ a \code{DBIConnection} object. }
}
\description{
  Not yet implemented.
}
\keyword{internal}

