\name{dbConnect,SQLiteDriver-method}
\alias{SQLITE_RO}
\alias{SQLITE_RW}
\alias{SQLITE_RWC}
\alias{dbConnect,SQLiteConnection-method}
\alias{dbConnect,SQLiteDriver-method}
\alias{dbConnect,character-method}
\alias{sqliteNewConnection}
\title{Create an SQLite connection.}
\usage{
  \S4method{dbConnect}{SQLiteDriver}(drv, ...)

  \S4method{dbConnect}{character}(drv, ...)

  \S4method{dbConnect}{SQLiteConnection}(drv, ...)

  sqliteNewConnection(drv, dbname = "",
    loadable.extensions = TRUE, cache_size = NULL,
    synchronous = 0, flags = NULL, vfs = NULL)
}
\arguments{
  \item{drv}{An objected generated by \code{\link{SQLite}},
  or an existing \code{\linkS4class{SQLiteConnection}}. If
  an connection, the connection will be cloned.}

  \item{dbname}{A path to a SQLite database. If the path
  does not exist, it will be created an initialied with an
  empty database.}

  \item{conn}{an \code{SQLiteConnection} object as produced
  by \code{dbConnect}.}

  \item{...}{Arguments passed on to
  \code{sqliteNewConnection}}

  \item{cache_size}{a positive integer to change the
  maximum number of disk pages that SQLite holds in memory
  (SQLite's default is 2000 pages).}

  \item{synchronous}{Possible values for \code{synchronous}
  are 0 (the default), 1, or 2 or the corresponding strings
  "OFF", "NORMAL", or "FULL".  Users have reported
  significant speed ups using \code{sychronous=0}, and the
  SQLite documentation itself implies considerable improved
  performance at the very modest risk of database
  corruption in the unlikely case of the operating system
  (\emph{not} the R application) crashing. See the SQLite
  documentation for the full details of this
  \code{PRAGMA}.}

  \item{flags}{\code{SQLITE_RWC}: open the database in
  read/write mode and create the database file if it does
  not already exist; \code{SQLITE_RW}: open the database in
  read/write mode. Raise an error if the file does not
  already exist; \code{SQLITE_RO}: open the database in
  read only mode.  Raise an error if the file does not
  already exist}
}
\description{
  Create an SQLite connection.
}

