\docType{methods}
\name{dbExistsTable,SQLiteConnection,character-method}
\alias{dbExistsTable,SQLiteConnection,character-method}
\title{Does the table exist?}
\usage{
\S4method{dbExistsTable}{SQLiteConnection,character}(conn, name, ...)
}
\arguments{
  \item{conn}{An existing
  \code{\linkS4class{SQLiteConnection}}}

  \item{name}{character vector of length 1 giving name of
  table}

  \item{...}{Ignored. Included for compatibility with
  generic.}
}
\description{
  Does the table exist?
}

