% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{dbExistsTable,SQLiteConnection,character-method}
\alias{dbExistsTable,SQLiteConnection,character-method}
\title{Does the table exist?}
\usage{
\S4method{dbExistsTable}{SQLiteConnection,character}(conn, name)
}
\arguments{
\item{conn}{An existing \code{\linkS4class{SQLiteConnection}}}

\item{name}{String, name of table. Match is case insensitive.}
}
\description{
Does the table exist?
}

