% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\docType{methods}
\name{dbExistsTable,SQLiteConnection,character-method}
\alias{dbExistsTable,SQLiteConnection,character-method}
\title{Does the table exist? See also \code{\link[DBI:dbExistsTable]{DBI::dbExistsTable()}} for the generic.}
\usage{
\S4method{dbExistsTable}{SQLiteConnection,character}(conn, name, ...)
}
\arguments{
\item{conn}{An existing \code{\linkS4class{SQLiteConnection}}}

\item{name}{String, name of table. Match is case insensitive.}

\item{...}{Needed for compatibility with generic. Otherwise ignored.}
}
\description{
Does the table exist? See also \code{\link[DBI:dbExistsTable]{DBI::dbExistsTable()}} for the generic.
}
\examples{
library(DBI)
db <- RSQLite::datasetsDb()

dbExistsTable(db, "mtcars")
dbExistsTable(db, "nonexistingtable")

dbDisconnect(db)
}

