% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{dbGetException,SQLiteConnection-method}
\alias{dbGetException,SQLiteConnection-method}
\title{Get the last exception from the connection.}
\usage{
\S4method{dbGetException}{SQLiteConnection}(conn)
}
\arguments{
\item{conn}{an object of class \code{\linkS4class{SQLiteConnection}}}

\item{...}{Ignored. Needed for compatiblity with generic.}
}
\description{
Get the last exception from the connection.
}

