% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{methods}
\name{dbGetInfo}
\alias{dbGetInfo}
\alias{dbGetInfo,SQLiteDriver-method}
\alias{dbGetInfo,SQLiteConnection-method}
\title{Get metadata about a database object}
\usage{
\S4method{dbGetInfo}{SQLiteDriver}(dbObj, ...)

\S4method{dbGetInfo}{SQLiteConnection}(dbObj, ...)
}
\arguments{
\item{dbObj}{An object of class \code{\linkS4class{SQLiteDriver}},
\code{\linkS4class{SQLiteConnection}} or
\code{\linkS4class{SQLiteResult}}}
}
\description{
Deprecated. Please use individual functions.
}
\keyword{internal}
