% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\docType{methods}
\name{dbListFields,SQLiteConnection,character-method}
\alias{dbListFields,SQLiteConnection,character-method}
\title{List fields in specified table}
\usage{
\S4method{dbListFields}{SQLiteConnection,character}(conn, name, ...)
}
\arguments{
\item{conn}{An existing \code{\linkS4class{SQLiteConnection}}}

\item{name}{a length 1 character vector giving the name of a table.}

\item{...}{Needed for compatibility with generic. Otherwise ignored.}
}
\description{
List fields in specified table
}
\examples{
con <- dbConnect(SQLite())
dbWriteTable(con, "iris", iris)
dbListFields(con, "iris")
dbDisconnect(con)
}

