% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{methods}
\name{dbListResults,SQLiteConnection-method}
\alias{dbListResults,SQLiteConnection-method}
\title{dbListResults}
\usage{
\S4method{dbListResults}{SQLiteConnection}(conn, ...)
}
\description{
DEPRECATED
}
\keyword{internal}

