\docType{methods}
\name{dbListResults,SQLiteConnection-method}
\alias{dbListResults,SQLiteConnection-method}
\title{List available SQLite result sets.}
\usage{
\S4method{dbListResults}{SQLiteConnection}(conn, ...)
}
\arguments{
  \item{conn}{An existing
  \code{\linkS4class{SQLiteConnection}}}

  \item{...}{Ignored. Included for compatibility with
  generic.}
}
\description{
  List available SQLite result sets.
}

