\name{dbListTables,SQLiteConnection-method}
\alias{dbListTables,SQLiteConnection-method}
\title{List available SQLite tables.}
\usage{
  \S4method{dbListTables}{SQLiteConnection}(conn, ...)
}
\arguments{
  \item{conn}{An existing
  \code{\linkS4class{SQLiteConnection}}}

  \item{...}{Ignored. Included for compatibility with
  generic.}
}
\description{
  List available SQLite tables.
}

