% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{dbSendPreparedQuery}
\alias{dbSendPreparedQuery}
\alias{dbGetPreparedQuery}
\title{Generics for getting and sending prepared queries}
\usage{
dbSendPreparedQuery(conn, statement, bind.data, ...)

dbGetPreparedQuery(conn, statement, bind.data, ...)
}
\arguments{
\item{conn}{A \code{DBIConnection} object.}

\item{statement}{A SQL string}

\item{bind.data}{A data frame}

\item{...}{Other arguments used by methods}
}
\description{
Generics for getting and sending prepared queries
}

