% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{dbUnloadDriver,SQLiteDriver-method}
\alias{dbUnloadDriver,SQLiteDriver-method}
\title{Unload SQLite driver.}
\usage{
\S4method{dbUnloadDriver}{SQLiteDriver}(drv, ...)
}
\arguments{
\item{drv}{Object created by \code{\link{SQLite}}}

\item{...}{Ignored. Needed for compatibility with generic.}
}
\value{
A logical indicating whether the operation succeeded or not.
}
\description{
Unload SQLite driver.
}
\examples{
db <- SQLite()
dbUnloadDriver(db)
}

