% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{methods}
\name{fetch,SQLiteResult-method}
\alias{fetch,SQLiteResult-method}
\title{Fetch}
\usage{
\S4method{fetch}{SQLiteResult}(res, n = -1, ...)
}
\description{
A shortcut for \code{\link[DBI]{dbFetch}(res, n = n, row.names = FALSE)},
kept for compatibility reasons.
}
\keyword{internal}
