% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbGetPreparedQuery.R,
%   R/dbGetPreparedQuery_SQLiteConnection_character_data.frame.R,
%   R/dbSendPreparedQuery.R,
%   R/dbSendPreparedQuery_SQLiteConnection_character_data.frame.R,
%   R/deprecated.R
\name{dbGetPreparedQuery}
\alias{dbGetPreparedQuery}
\alias{dbGetPreparedQuery_SQLiteConnection_character_data.frame}
\alias{dbGetPreparedQuery,SQLiteConnection,character,data.frame-method}
\alias{dbSendPreparedQuery}
\alias{dbSendPreparedQuery_SQLiteConnection_character_data.frame}
\alias{dbSendPreparedQuery,SQLiteConnection,character,data.frame-method}
\alias{query-dep}
\title{Deprecated querying tools}
\usage{
dbGetPreparedQuery(conn, statement, bind.data, ...)

\S4method{dbGetPreparedQuery}{SQLiteConnection,character,data.frame}(conn, statement, bind.data, ...)

dbSendPreparedQuery(conn, statement, bind.data, ...)

\S4method{dbSendPreparedQuery}{SQLiteConnection,character,data.frame}(conn, statement, bind.data, ...)
}
\arguments{
\item{conn}{A \code{DBIConnection} object.}

\item{statement}{A SQL string}

\item{bind.data}{A data frame of data to be bound.}

\item{...}{Other arguments used by methods}
}
\description{
These functions have been deprecated. Please switch to using
\code{\link[DBI:dbSendQuery]{DBI::dbSendQuery()}}/\code{\link[DBI:dbGetQuery]{DBI::dbGetQuery()}} with the \code{params} argument
or with calling \code{\link[DBI:dbBind]{DBI::dbBind()}} instead.
}
\keyword{internal}
