% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{methods}
\name{dbSendPreparedQuery}
\alias{dbSendPreparedQuery}
\alias{dbGetPreparedQuery}
\alias{query-dep}
\alias{dbSendPreparedQuery,SQLiteConnection,character,data.frame-method}
\alias{dbGetPreparedQuery,SQLiteConnection,character,data.frame-method}
\title{Deprecated querying tools}
\usage{
dbSendPreparedQuery(conn, statement, bind.data, ...)

dbGetPreparedQuery(conn, statement, bind.data, ...)

\S4method{dbSendPreparedQuery}{SQLiteConnection,character,data.frame}(conn,
  statement, bind.data, ...)

\S4method{dbGetPreparedQuery}{SQLiteConnection,character,data.frame}(conn,
  statement, bind.data, ...)
}
\arguments{
\item{conn}{A \code{DBIConnection} object.}

\item{statement}{A SQL string}

\item{bind.data}{A data frame}

\item{...}{Other arguments used by methods}

\item{bind.data}{A data frame of data to be bound.}
}
\description{
These functions have been deprecated. Please switch to using
\code{\link[=dbSendQuery]{dbSendQuery()}}/\code{\link[=dbGetQuery]{dbGetQuery()}} with the \code{params} argument
or with calling \code{\link[=dbBind]{dbBind()}} instead.
}
\keyword{internal}
