% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rsqliteVersion}
\alias{rsqliteVersion}
\title{RSQLite version}
\usage{
rsqliteVersion()
}
\value{
A character vector containing header and library versions of
  RSQLite.
}
\description{
RSQLite version
}

