% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{rsqliteVersion}
\alias{rsqliteVersion}
\title{RSQLite version}
\usage{
rsqliteVersion()
}
\value{
A character vector containing header and library versions of
RSQLite.
}
\description{
Return the version of RSQLite.
}
\examples{
RSQLite::rsqliteVersion()
}
