% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{sqliteBuildTableDefinition}
\alias{sqliteBuildTableDefinition}
\alias{dbBuildTableDefinition}
\title{Build the SQL CREATE TABLE definition as a string}
\usage{
sqliteBuildTableDefinition(
  con,
  name,
  value,
  field.types = NULL,
  row.names = pkgconfig::get_config("RSQLite::row.names.query", FALSE)
)
}
\arguments{
\item{con}{A database connection.}

\item{name}{Name of the new SQL table}

\item{value}{A data.frame, for which we want to create a table.}

\item{field.types}{Optional, named character vector of the types for each field in \code{value}}

\item{row.names}{Logical.
 Should row.name of \code{value} be exported as a \code{row_names} field?
 Default is \code{TRUE}}
}
\value{
An SQL string
}
\description{
The output SQL statement is a simple \verb{CREATE TABLE} suitable for \code{dbGetQuery}
}
\keyword{internal}
