% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sqliteBuildTableDefinition}
\alias{dbBuildTableDefinition}
\alias{sqliteBuildTableDefinition}
\title{Build the SQL CREATE TABLE definition as a string}
\usage{
sqliteBuildTableDefinition(con, name, value, field.types = NULL,
  row.names = NA, ...)
}
\arguments{
\item{name}{Name of the new SQL table}

\item{value}{A data.frame, for which we want to create a table.}

\item{field.types}{Optional, named character vector of the types for each
field in \code{value}}

\item{row.names}{Logical. Should row.name of \code{value} be exported as a
\code{row\_names} field? Default is \code{TRUE}}

\item{...}{Ignored. Reserved for future use.}

\item{conn}{A database connection.}
}
\value{
An SQL string
}
\description{
The output SQL statement is a simple \code{CREATE TABLE} with suitable for
\code{dbGetQuery}
}
\keyword{internal}

