% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLiteConnection.R, R/table.R
\docType{class}
\name{SQLiteConnection-class}
\alias{SQLiteConnection-class}
\alias{dbQuoteIdentifier,SQLiteConnection,character-method}
\alias{dbQuoteIdentifier,SQLiteConnection,SQL-method}
\alias{show,SQLiteConnection-method}
\alias{dbIsValid,SQLiteConnection-method}
\alias{dbGetException,SQLiteConnection-method}
\alias{dbDataType,SQLiteConnection-method}
\title{Class SQLiteConnection}
\usage{
\S4method{dbQuoteIdentifier}{SQLiteConnection,character}(conn, x, ...)

\S4method{dbQuoteIdentifier}{SQLiteConnection,SQL}(conn, x, ...)

\S4method{show}{SQLiteConnection}(object)

\S4method{dbIsValid}{SQLiteConnection}(dbObj, ...)

\S4method{dbGetException}{SQLiteConnection}(conn, ...)

\S4method{dbDataType}{SQLiteConnection}(dbObj, obj, ...)
}
\description{
\code{SQLiteConnection} objects are usually created by
\code{\link[DBI:dbConnect]{DBI::dbConnect()}}.
}
\keyword{internal}
