% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Connection.R
\docType{class}
\name{SQLiteConnection-class}
\alias{SQLiteConnection-class}
\title{Class SQLiteConnection.}
\description{
\code{SQLiteConnection} objects are usually created by
\code{\link[DBI]{dbConnect}}
}
\examples{
con <- dbConnect(SQLite(), dbname = tempfile())
dbDisconnect(con)
}

