% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLiteConnection.R
\docType{class}
\name{SQLiteConnection-class}
\alias{SQLiteConnection-class}
\alias{dbGetException,SQLiteConnection-method}
\alias{dbIsValid,SQLiteConnection-method}
\alias{show,SQLiteConnection-method}
\title{Class SQLiteConnection.}
\usage{
\S4method{show}{SQLiteConnection}(object)

\S4method{dbIsValid}{SQLiteConnection}(dbObj, ...)

\S4method{dbGetException}{SQLiteConnection}(conn, ...)
}
\description{
\code{SQLiteConnection} objects are usually created by
\code{\link[DBI]{dbConnect}}.
}
\keyword{internal}

