% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLiteDriver.R, R/dbDataType_SQLiteDriver.R,
%   R/dbGetInfo_SQLiteDriver.R, R/dbIsValid_SQLiteDriver.R,
%   R/dbUnloadDriver_SQLiteDriver.R
\docType{class}
\name{SQLiteDriver-class}
\alias{SQLiteDriver-class}
\alias{dbDataType_SQLiteDriver}
\alias{dbDataType,SQLiteDriver-method}
\alias{dbGetInfo_SQLiteDriver}
\alias{dbGetInfo,SQLiteDriver-method}
\alias{dbIsValid_SQLiteDriver}
\alias{dbIsValid,SQLiteDriver-method}
\alias{dbUnloadDriver_SQLiteDriver}
\alias{dbUnloadDriver,SQLiteDriver-method}
\title{Class SQLiteDriver (and methods)}
\usage{
\S4method{dbDataType}{SQLiteDriver}(dbObj, obj, ..., extended_types = FALSE)

\S4method{dbGetInfo}{SQLiteDriver}(dbObj, ...)

\S4method{dbIsValid}{SQLiteDriver}(dbObj, ...)

\S4method{dbUnloadDriver}{SQLiteDriver}(drv, ...)
}
\description{
SQLiteDriver objects are created by \code{\link[=SQLite]{SQLite()}},
and used to select the correct method in \code{\link[DBI:dbConnect]{dbConnect()}}.
They are a superclass of the \link[DBI:DBIDriver-class]{DBIDriver} class,
and used purely for dispatch.
The "Usage" section lists the class methods overridden by \pkg{RSQLite}.
The \code{\link[DBI:dbUnloadDriver]{dbUnloadDriver()}} method is a null-op.
}
\keyword{internal}
