% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLiteDriver.R, R/table.R
\docType{class}
\name{SQLiteDriver-class}
\alias{SQLiteDriver-class}
\alias{dbIsValid,SQLiteDriver-method}
\alias{dbUnloadDriver,SQLiteDriver-method}
\alias{dbDataType,SQLiteDriver-method}
\title{Class SQLiteDriver (and methods)}
\usage{
\S4method{dbIsValid}{SQLiteDriver}(dbObj, ...)

\S4method{dbUnloadDriver}{SQLiteDriver}(drv, ...)

\S4method{dbDataType}{SQLiteDriver}(dbObj, obj, ...)
}
\description{
The SQLiteDriver, which is used to select the correct method in
\code{\link[=dbConnect]{dbConnect()}}. See more details in \code{\link[=SQLite]{SQLite()}}.
It is used purely for dispatch and \code{\link[=dbUnloadDriver]{dbUnloadDriver()}} is a null-op.
}
\keyword{internal}
