% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Driver.R
\docType{class}
\name{SQLiteDriver-class}
\alias{SQLiteDriver-class}
\alias{show,SQLiteDriver-method}
\title{Class SQLiteDriver (and methods).}
\usage{
\S4method{show}{SQLiteDriver}(object)
}
\description{
The SQLiteDriver, which is used to select the correct method in
\code{dbConnect()}. See more details in \code{\link{SQLite}}
}
\keyword{internal}

