% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLiteDriver.R
\docType{class}
\name{SQLiteDriver-class}
\alias{SQLiteDriver-class}
\alias{dbDataType,SQLiteDriver-method}
\alias{dbIsValid,SQLiteDriver-method}
\alias{dbUnloadDriver,SQLiteDriver-method}
\title{Class SQLiteDriver (and methods)}
\usage{
\S4method{dbDataType}{SQLiteDriver}(dbObj, obj, ...)

\S4method{dbIsValid}{SQLiteDriver}(dbObj, ...)

\S4method{dbUnloadDriver}{SQLiteDriver}(drv, ...)
}
\description{
SQLiteDriver objects are created by \code{\link[=SQLite]{SQLite()}}, and used to select the
correct method in \code{\link[=dbConnect]{dbConnect()}}.
They are a superclass of the \linkS4class{DBIDriver} class,
and used purely for dispatch.
The "Usage" section lists the class methods overridden by \pkg{RSQLite}.
The \code{\link[=dbUnloadDriver]{dbUnloadDriver()}} method is a null-op.
}
\keyword{internal}
