% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{sqliteQuickColumn}
\alias{sqliteQuickColumn}
\title{Return an entire column from a SQLite database}
\usage{
sqliteQuickColumn(con, table, column)
}
\description{
DEPRECATED. Please use dbReadTable instead.
}
\keyword{internal}

