% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{sqliteQuickColumn}
\alias{sqliteQuickColumn}
\title{Return an entire column from a SQLite database}
\usage{
sqliteQuickColumn(con, table, column)
}
\description{
A shortcut for
\code{\link[DBI]{dbReadTable}(con, table, select.cols = column, row.names = FALSE)[[1]]},
kept for compatibility reasons.
}
\keyword{internal}
