% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpvfs.R
\name{sqliteRemote}
\alias{sqliteRemote}
\title{Convenience helper for opening remote SQLite databases over HTTP/HTTPS}
\usage{
sqliteRemote(url, immutable = TRUE, ...)
}
\arguments{
\item{url}{Remote HTTP/HTTPS URL to a SQLite database file.}

\item{immutable}{Logical, append \code{immutable=1} to the URI.}

\item{...}{Passed through to \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}
}
\value{
A \link[=SQLiteConnection-class]{SQLiteConnection} object.
}
\description{
lifecycle::badge("experimental")

Constructs a URI filename and sets the appropriate VFS and immutable flags.
Requires the optional http extension to be available; see \code{\link[=sqliteHasHttpVFS]{sqliteHasHttpVFS()}}.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
if (sqliteHasHttpVFS()) {
  con <- sqliteRemote("https://example.org/db.sqlite")
  dbGetQuery(con, "SELECT name FROM sqlite_master WHERE type='table'")
  dbDisconnect(con)
}
\dontshow{\}) # examplesIf}
}
