% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SQLiteResult.R
\docType{class}
\name{SQLiteResult-class}
\alias{SQLiteResult-class}
\alias{dbIsValid,SQLiteResult-method}
\title{Class SQLiteResult}
\usage{
\S4method{dbIsValid}{SQLiteResult}(dbObj)
}
\description{
SQLite's query results class.  This classes encapsulates the result of an
SQL statement (either \code{select} or not).
}
\keyword{internal}

