\docType{class}
\name{SQLiteResult-class}
\alias{SQLiteResult-class}
\title{Class SQLiteResult}
\description{
  SQLite's query results class.  This classes encapsulates
  the result of an SQL statement (either \code{select} or
  not).
}
\section{Generators}{
  The main generator is \code{\link[DBI]{dbSendQuery}}.
}

