% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Result.R
\docType{class}
\name{SQLiteResult-class}
\alias{SQLiteResult-class}
\title{Class SQLiteResult}
\description{
SQLite's query results class.  This classes encapsulates the result of an
SQL statement (either \code{select} or not).
}

