% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLiteResult.R, R/query.R
\docType{class}
\name{SQLiteResult-class}
\alias{SQLiteResult-class}
\alias{dbIsValid,SQLiteResult-method}
\alias{dbBind,SQLiteResult-method}
\alias{dbFetch,SQLiteResult-method}
\alias{dbClearResult,SQLiteResult-method}
\title{Class SQLiteResult}
\usage{
\S4method{dbIsValid}{SQLiteResult}(dbObj, ...)

\S4method{dbBind}{SQLiteResult}(res, params, ...)

\S4method{dbFetch}{SQLiteResult}(res, n = -1, ..., row.names = NA)

\S4method{dbClearResult}{SQLiteResult}(res, ...)
}
\description{
SQLite's query results class.  This classes encapsulates the result of an
SQL statement (either \code{SELECT} or not).
}
\seealso{
The corresponding generic functions
\code{\link[DBI:dbFetch]{DBI::dbFetch()}}, \code{\link[DBI:dbClearResult]{DBI::dbClearResult()}}, and \code{\link[DBI:dbBind]{DBI::dbBind()}}.
}
\keyword{internal}
