% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,SQLiteConnection-method}
\alias{summary,SQLiteDriver-method}
\alias{summary,SQLiteResult-method}
\title{Summary methods}
\usage{
\S4method{summary}{SQLiteDriver}(object)

\S4method{summary}{SQLiteConnection}(object)

\S4method{summary}{SQLiteResult}(object)
}
\arguments{
\item{verbose}{Show extra information.}
}
\description{
Summary methods
}
\examples{
summary(SQLite())

con <- dbConnect(SQLite())
summary(con)

dbWriteTable(con, "mtcars", mtcars)
rs <- dbSendQuery(con, "SELECT * FROM mtcars")
summary(rs)

dbClearResults(rs)
dbDisconnect(con)
}
\keyword{internal}

