% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Summary.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,SQLiteConnection-method}
\alias{summary,SQLiteResult-method}
\title{Summary methods}
\usage{
\S4method{summary}{SQLiteConnection}(object)

\S4method{summary}{SQLiteResult}(object)
}
\description{
Summary methods
}
\examples{
summary(SQLite())

con <- dbConnect(SQLite())
summary(con)

dbWriteTable(con, "mtcars", mtcars)
rs <- dbSendQuery(con, "SELECT * FROM mtcars")
summary(rs)

dbClearResult(rs)
dbDisconnect(con)
}
\keyword{internal}

