\docType{methods}
\name{summary}
\alias{sqliteDescribeConnection}
\alias{sqliteDescribeDriver}
\alias{sqliteDescribeResult}
\alias{summary}
\alias{summary,SQLiteConnection-method}
\alias{summary,SQLiteDriver-method}
\alias{summary,SQLiteResult-method}
\title{Summary methods}
\usage{
\S4method{summary}{SQLiteDriver}(object, ...)

sqliteDescribeDriver(obj, verbose = FALSE, ...)

\S4method{summary}{SQLiteConnection}(object, ...)

sqliteDescribeConnection(obj, verbose = FALSE, ...)

\S4method{summary}{SQLiteResult}(object, ...)

sqliteDescribeResult(obj, verbose = FALSE, ...)
}
\description{
Summary methods
}
\keyword{internal}

